% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_bits}
\alias{count_bits}
\title{Count the number of bits set to 1 in a positive natural number}
\usage{
count_bits(x)
}
\arguments{
\item{x}{the objective number}
}
\value{
the number of bits
}
\description{
Count the number of bits set to 1 in a positive natural number
}
