% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso_main.R
\name{debug_foo}
\alias{debug_foo}
\title{Just a debug function to try out rcpp stuff}
\usage{
debug_foo()
}
\value{
whatever you want to return in the testing
}
\description{
Modify 'debug_cpp' to test behaviours and interactions down in C++
}
