% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causality_list.R
\name{natCauslist}
\alias{natCauslist}
\title{R6 class that defines causal lists in the PSO}
\arguments{
\item{ordering}{a vector with the names of the nodes in t_0}
}
\value{
A new 'natCauslist' object

the ordering with the names cropped
}
\description{
Constructor of the 'natCauslist' class

If the names of the nodes have "_t_0" appended at the end, remove it
}
\details{
The causal lists will be the base of the positions and the velocities
in the pso part of the algorithm. They will not have the same structure
as their binary counterparts, but their class skeleton will serve as a
base.
}
\section{Fields}{

\describe{
\item{\code{cl}}{List of causal units}

\item{\code{ordering}}{String vector defining the order of the nodes in t_0}

\item{\code{ordering}}{String vector defining the order of the nodes without the appended "_t_0"}
}}

