% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{natPosition}
\alias{natPosition}
\title{R6 class that defines DBNs as causality lists}
\arguments{
\item{nodes}{A list with the names of the nodes in t_0 in the network
If its not null, a random position will be generated for those nodes.}

\item{vl}{a natVelocity object}

\item{ps}{a natPosition object
return the natVelocity that gets the other position to this one}

\item{net}{a dbn object}

\item{n_vars}{the number of variables in t_0}

\item{max_size}{the maximum size of the DBN}

\item{p}{the parameter of the truncated geometric sampler. If lesser or
equal to 0, a uniform distribution will be used instead.}
}
\value{
A new 'natPosition' object

a dbn object

the ordering of the nodes in t_0

a causlist object

a random position

the number of arcs
}
\description{
Constructor of the 'natPosition' class

Translate the causality list into a DBN network

Uses this object private causality list and transforms it into a DBN.

Add a velocity to the position

Given a natVelocity object, add it to the current position.

Given another position, returns the velocity that gets it to this
position.

Return the static node ordering

This function takes as input a dbn and return the node ordering of the
variables inside a timeslice. This ordering is needed to understand a
causal list.

Translate a DBN into a causality list

This function takes as input a network from a DBN and transforms the 
structure into a causality list if it is a valid DBN. Valid DBNs have only
inter-timeslice edges and only allow variables in t_0 to have parents.

Generates a random position

This function takes as input the number of variables, the maximum size
and the parameter p and returns a random position with arcs 
sampled either from the uniform distribution or from a truncated 
geometric distribution. Much faster than the binary implementation with
lists of lists and random bn generation into translation.

Recount the number of arcs in the cl
}
\details{
A causality list has a list with causal units, a size representing the
Markovian order of the network and a specific node ordering.
}
\section{Fields}{

\describe{
\item{\code{n_arcs}}{Number of arcs in the network}

\item{\code{max_size}}{Maximum number of timeslices of the DBN}

\item{\code{p}}{Parameter of the sampling truncated geometric distribution}

\item{\code{nodes}}{Names of the nodes in the network}
}}

