% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso_main.R
\name{debug_foo}
\alias{debug_foo}
\title{Just a debug function to try out rcpp stuff}
\usage{
debug_foo(x)
}
\arguments{
\item{x}{some parameter}
}
\value{
whatever you want to return in the testing
}
\description{
Modify 'debug_cpp' to test behaviours and interactions down in C++
Currently: testing initialization times in both R and C++
Results: equivalent times, O(n) as it should be. Will leave the R initialization.
}
